+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Invoice : Fix multiple invoices not ventilating when generated from sale order with advance payment 
 - BANK PAYMENT : fix ICS number being linked to EBICS
 - ICalendar : Resolve synchronisation issue & Mange synchronisation select.
+- Batch payroll preparation generation : corrected batch error no session.
 
 ## [5.1.15] - 2020-03-23
 ## Improvements